order using bin packing but at a different layer
